% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-openedge.R
\name{backend-openedge}
\alias{backend-openedge}
\title{Backend: OpenEdge}
\description{
Based on Microsoft SQL Server implementation. Subclasses the simulated DBI
for Microsoft SQL.
\itemize{
\item \code{SELECT} uses \code{TOP} not \code{LIMIT}
\item Analysis uses \verb{UPDATE STATISTICS} rather than \code{ANALYZE}
}
}
