% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressDataDirectConnect4R-package.R
\docType{package}
\name{ProgressDataDirectConnect4R-package}
\alias{ProgressDataDirectConnect4R}
\alias{ProgressDataDirectConnect4R-package}
\title{ProgressDataDirectConnect4R: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
