% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{unzip.jars}
\alias{unzip.jars}
\title{Downloads and unzips DataDirect Connect for JDBC from Progress}
\usage{
unzip.jars(exdir = tempdir())
}
\arguments{
\item{exdir}{Directory to extract Jar files to; created if necessary}
}
\description{
Downloads the 45-megabyte DataDirect JDBC evaluation. Skips the name and
e-mail requirement. Operates interactively by design. Extracts and caches the
Java archives ready for installation. Run periodically to update.
}
\details{
Makes you wonder: does caching satisfy the licensing requirement?
}
